리뷰반영, handler분리
