+++ cypress/e2e/step1.cy.js
@@ -1,3 +1,5 @@
+import { CHARGE, PRODUCT } from '../../js/constants/message';
+
 const SELECTOR = {
   TAB_PRODUCT_MENU: '#product-manage-menu',
   TAB_CHARGE_MENU: '#vending-machine-manage-menu',
@@ -23,13 +25,13 @@ const SELECTOR = {
   COIN_QUANTITY_10: '#vending-machine-coin-10-quantity',
 };
 
-const MESSAGE = {
-  PRODUCT_MIN_PRICE: '최소 가격은 100원입니다.',
-  PRODUCT_PRICE_UNIT: '가격 단위는 10원입니다.',
-  PRODUCT_MIN_QUANTITY: '최소 수량은 1개입니다.',
+const addProduct = (value) => {
+  const [name, price, amount] = value;
+  cy.get('#product-name-input').type(name);
+  cy.get('#product-price-input').type(price);
+  cy.get('#product-quantity-input').type(amount);
 
-  CHARGE_AMOUNT_MIN: '최소 충전 금액은 100원입니다.',
-  CHARGE_AMOUNT_UNIT: '충전 단위는 10원입니다.',
+  cy.get('#product-add-button').click();
 };
 
 beforeEach(() => {
@@ -65,11 +67,11 @@ describe('상품 관리를 한다', () => {
 
     cy.get(SELECTOR.PRODUCT_PRICE).type('99');
     cy.get(SELECTOR.PRODUCT_ADD_BUTTON).click();
-    cy.alertMessage(SELECTOR.PRODUCT_ADD_BUTTON, MESSAGE.PRODUCT_MIN_PRICE);
+    cy.alertMessage(SELECTOR.PRODUCT_ADD_BUTTON, PRODUCT.MIN_PRICE);
 
     cy.get(SELECTOR.PRODUCT_PRICE).type('111');
     cy.get(SELECTOR.PRODUCT_ADD_BUTTON).click();
-    cy.alertMessage(SELECTOR.PRODUCT_ADD_BUTTON, MESSAGE.PRODUCT_PRICE_UNIT);
+    cy.alertMessage(SELECTOR.PRODUCT_ADD_BUTTON, PRODUCT.PRICE_UNIT);
 
     cy.get(SELECTOR.PRODUCT_PRICE).type('200');
     cy.get(SELECTOR.PRODUCT_ADD_BUTTON).click();
@@ -83,11 +85,11 @@ describe('상품 관리를 한다', () => {
     cy.get(SELECTOR.PRODUCT_QUANTITY).should('have.not.value', ' ');
   });
   it('최소 수량은 1개이다', () => {
-    cy.addProduct(['콜라', '200', '0']);
-    cy.alertMessage(SELECTOR.PRODUCT_ADD_BUTTON, MESSAGE.PRODUCT_MIN_QUANTITY);
+    addProduct(['콜라', '200', '0']);
+    cy.alertMessage(SELECTOR.PRODUCT_ADD_BUTTON, PRODUCT.MIN_QUANTITY);
   });
   it('추가하기를 클릭하면 form은 초기화된다', () => {
-    cy.addProduct(['콜라', '200', '1']);
+    addProduct(['콜라', '200', '1']);
 
     cy.get(SELECTOR.PRODUCT_ADD_BUTTON).click();
 
@@ -96,7 +98,7 @@ describe('상품 관리를 한다', () => {
     cy.get(SELECTOR.PRODUCT_QUANTITY).should('have.value', '');
   });
   it('추가하기를 클릭하면 상품 목록에 표기된다', () => {
-    cy.addProduct(['콜라', '200', '1']);
+    addProduct(['콜라', '200', '1']);
 
     cy.get(SELECTOR.PRODUCT_ADD_BUTTON)
       .click()
@@ -105,13 +107,12 @@ describe('상품 관리를 한다', () => {
       });
   });
   it('같은 상품명의 데이터를 추가하면 대체된다', () => {
-    cy.addProduct(['콜라', '200', '1']);
-    cy.addProduct(['콜라', '2000', '5']).then(() => {
-      cy.get(SELECTOR.PRODUCT_LIST).should('have.length', 1);
-      cy.get(SELECTOR.PRODUCT_LIST).should('contain', '콜라');
-      cy.get(SELECTOR.PRODUCT_LIST).should('contain', '2000');
-      cy.get(SELECTOR.PRODUCT_LIST).should('contain', '5');
-    });
+    addProduct(['콜라', '200', '1']);
+    addProduct(['콜라', '2000', '5']);
+    cy.get(SELECTOR.PRODUCT_LIST).should('have.length', 1);
+    cy.get(SELECTOR.PRODUCT_LIST).should('contain', '콜라');
+    cy.get(SELECTOR.PRODUCT_LIST).should('contain', '2000');
+    cy.get(SELECTOR.PRODUCT_LIST).should('contain', '5');
   });
 });
 
@@ -136,7 +137,7 @@ describe('잔돈 충전을 한다.', () => {
     cy.get(SELECTOR.COIN_QUANTITY_10).should('exist');
     cy.get(SELECTOR.COIN_QUANTITY_10).should('contain', '0개');
   });
-  it('충전할 금액을 입력해 충전할 수 있다', () => {
+  it('충전을 할 수 있다', () => {
     cy.get(SELECTOR.CHARGE_INPUT).should('exist');
     cy.get(SELECTOR.CHARGE_INPUT).type('1000원');
 
@@ -148,11 +149,11 @@ describe('잔돈 충전을 한다.', () => {
   });
   it('충전할 금액이 최소금액보다 작으면 alert가 발생한다', () => {
     cy.get(SELECTOR.CHARGE_INPUT).type('90원');
-    cy.alertMessage(SELECTOR.CHARGE_BUTTON, MESSAGE.CHARGE_AMOUNT_MIN);
+    cy.alertMessage(SELECTOR.CHARGE_BUTTON, CHARGE.MIN_AMOUNT);
   });
   it('충전할 금액이 10원으로 나누어 떨어지지 않으면 alert가 발생한다', () => {
     cy.get(SELECTOR.CHARGE_INPUT).type('111원');
-    cy.alertMessage(SELECTOR.CHARGE_BUTTON, MESSAGE.CHARGE_AMOUNT_UNIT);
+    cy.alertMessage(SELECTOR.CHARGE_BUTTON, CHARGE.AMOUNT_UNIT);
   });
   it('잔돈을 누적해서 충전할 수 있다', () => {
     cy.get(SELECTOR.CHARGE_INPUT).type('1000');
