+++ js/controllers/charge.js
@@ -1,6 +1,8 @@
-import { COINS } from '../constants/constant.js';
-import { vendingMachine } from '../models/VendingMachine.js';
+import { CHARGE, COINS } from '../constants/constant.js';
+import { vendingMachine } from '../index.js';
 import { getRandom } from '../utils/random.js';
+import { validate } from '../utils/validator.js';
+import { chargeAmountValidations } from '../validations/validation.js';
 import {
   clearChargeForm,
   renderCoins,
@@ -10,7 +12,7 @@ import {
 export const addCharge = (chargeAmount) => {
   const machineChargeItem = vendingMachine.getCharge();
   const amount = machineChargeItem.amount + chargeAmount;
-  const coins = getCoins(chargeAmount);
+  const coins = addCoins(chargeAmount);
 
   vendingMachine.setCharge({
     amount: amount,
@@ -23,14 +25,15 @@ export const addCharge = (chargeAmount) => {
   clearChargeForm();
 };
 
-export const getCoins = (amount) => {
+export const addCoins = (amount) => {
   const coinUnit = Object.values(COINS);
   const coins = vendingMachine.getCharge().coins;
 
   let idx = 1;
   for (const coin of coinUnit) {
-    const max = Math.floor(amount / coin);
-    const quantity = idx === coinUnit.length ? max : getRandom(0, max);
+    const coinMaxNumber = Math.floor(amount / coin);
+    const quantity =
+      idx === coinUnit.length ? coinMaxNumber : getRandom(0, coinMaxNumber);
 
     amount = amount - quantity * coin;
     coins[COINS[`COIN_${coin}`]] = coins[COINS[`COIN_${coin}`]] + quantity;
@@ -39,4 +42,15 @@ export const getCoins = (amount) => {
   }
 
   return coins;
+};
+
+export const handleFormChargeSubmit = (event) => {
+  event.preventDefault();
+  const chargeAmount = event.target.elements[CHARGE.AMOUNT].value;
+  try {
+    validate(chargeAmount, chargeAmountValidations);
+    addCharge(+chargeAmount);
+  } catch (err) {
+    alert(err.message);
+  }
 };
