+++ js/index.js
@@ -1,10 +1,12 @@
 import { SELECTOR } from './constants/constant.js';
 import { $ } from './utils/selector.js';
-import {
-  handFormChargeSubmit,
-  handleChangeTab,
-  handleFormProductSubmit,
-} from './handler.js';
+
+import { VendingMachine } from './models/VendingMachine.js';
+import { handleChangeTab } from './controllers/tab.js';
+import { handleFormProductSubmit } from './controllers/product.js';
+import { handleFormChargeSubmit } from './controllers/charge.js';
+
+export const vendingMachine = new VendingMachine();
 
 export const setProductEvent = () => {
   $(SELECTOR.PRODUCT_FORM).addEventListener('submit', (event) => {
@@ -14,7 +16,7 @@ export const setProductEvent = () => {
 
 export const setChargeEvent = () => {
   $(SELECTOR.CHARGE_FORM).addEventListener('submit', (event) => {
-    handFormChargeSubmit(event);
+    handleFormChargeSubmit(event);
   });
 };
